refactor/app Move App into its own module for better separation
