+++ safe_app/src/lib.rs
@@ -66,6 +66,7 @@ extern crate tokio_core;
 #[macro_use]
 extern crate unwrap;
 
+pub mod app;
 pub mod ffi;
 pub mod object_cache;
 
@@ -73,368 +74,5 @@ mod errors;
 #[cfg(test)]
 mod test_utils;
 
-use futures::{Future, future};
-use futures::stream::Stream;
-use futures::sync::mpsc as futures_mpsc;
-use maidsafe_utilities::serialisation::deserialise;
-use maidsafe_utilities::thread::{self, Joiner};
-use rust_sodium::crypto::hash::sha256;
-use rust_sodium::crypto::secretbox;
-use safe_core::{Client, ClientKeys, CoreMsg, CoreMsgTx, FutureExt, MDataInfo, NetworkEvent,
-                NetworkTx, event_loop, utils};
-use safe_core::ipc::{AccessContInfo, AppKeys, AuthGranted, Permission};
 pub use self::errors::*;
-use self::object_cache::ObjectCache;
-use std::cell::RefCell;
-use std::collections::{BTreeSet, HashMap};
-use std::rc::Rc;
-use std::sync::Mutex;
-use std::sync::mpsc as std_mpsc;
-use tokio_core::reactor::{Core, Handle};
-
-macro_rules! try_tx {
-    ($result:expr, $tx:ident) => {
-        match $result {
-            Ok(res) => res,
-            Err(e) => return unwrap!($tx.send(Err(AppError::from(e)))),
-        }
-    }
-}
-
-type AppFuture<T> = Future<Item = T, Error = AppError>;
-
-/// Handle to an application instance.
-pub struct App {
-    core_tx: Mutex<CoreMsgTx<AppContext>>,
-    _core_joiner: Joiner,
-}
-
-impl App {
-    /// Create unregistered app.
-    pub fn unregistered<N>(network_observer: N) -> Result<Self, AppError>
-        where N: FnMut(Result<NetworkEvent, AppError>) + Send + 'static
-    {
-        Self::new(network_observer, |el_h, core_tx, net_tx| {
-            let client = Client::unregistered(el_h, core_tx, net_tx)?;
-            let context = AppContext::unregistered();
-            Ok((client, context))
-        })
-    }
-
-    /// Create registered app.
-    pub fn registered<N>(app_id: String,
-                         auth_granted: AuthGranted,
-                         network_observer: N)
-                         -> Result<Self, AppError>
-        where N: FnMut(Result<NetworkEvent, AppError>) + Send + 'static
-    {
-        let AuthGranted {
-            app_keys: AppKeys { owner_key, enc_key, enc_pk, enc_sk, sign_pk, sign_sk },
-            access_container,
-            ..
-        } = auth_granted;
-
-        let client_keys = ClientKeys {
-            sign_pk: sign_pk,
-            sign_sk: sign_sk,
-            enc_pk: enc_pk,
-            enc_sk: enc_sk,
-        };
-
-        Self::new(network_observer, move |el_h, core_tx, net_tx| {
-            let client = Client::from_keys(client_keys, owner_key, el_h, core_tx, net_tx)?;
-            let context = AppContext::registered(app_id, enc_key, access_container);
-            Ok((client, context))
-        })
-    }
-
-    fn new<N, F>(mut network_observer: N, setup: F) -> Result<Self, AppError>
-        where N: FnMut(Result<NetworkEvent, AppError>) + Send + 'static,
-              F: FnOnce(Handle, CoreMsgTx<AppContext>, NetworkTx)
-                        -> Result<(Client, AppContext), AppError> + Send + 'static
-    {
-        let (tx, rx) = std_mpsc::sync_channel(0);
-
-        let joiner = thread::named("App Event Loop", move || {
-            let el = try_tx!(Core::new(), tx);
-            let el_h = el.handle();
-
-            let (core_tx, core_rx) = futures_mpsc::unbounded();
-            let (net_tx, net_rx) = futures_mpsc::unbounded();
-
-            el_h.spawn(net_rx.map(move |event| network_observer(Ok(event)))
-                .for_each(|_| Ok(())));
-
-            let core_tx_clone = core_tx.clone();
-
-            let (client, context) = try_tx!(setup(el_h, core_tx_clone, net_tx), tx);
-            unwrap!(tx.send(Ok(core_tx)));
-
-            event_loop::run(el, client, context, core_rx);
-        });
-
-        let core_tx = rx.recv()??;
-
-        Ok(App {
-            core_tx: Mutex::new(core_tx),
-            _core_joiner: joiner,
-        })
-    }
-
-    /// Send a message to app's event loop
-    pub fn send<F>(&self, f: F) -> Result<(), AppError>
-        where F: FnOnce(&Client, &AppContext) -> Option<Box<Future<Item=(), Error=()>>>
-                 + Send + 'static
-    {
-        let msg = CoreMsg::new(f);
-        let mut core_tx = unwrap!(self.core_tx.lock());
-        core_tx.send(msg).map_err(AppError::from)
-    }
-}
-
-impl Drop for App {
-    fn drop(&mut self) {
-        let mut core_tx = match self.core_tx.lock() {
-            Ok(core_tx) => core_tx,
-            Err(err) => {
-                info!("Unexpected error in drop: {:?}", err);
-                return;
-            }
-        };
-
-        let msg = CoreMsg::build_terminator();
-        if let Err(err) = core_tx.send(msg) {
-            info!("Unexpected error in drop: {:?}", err);
-        }
-    }
-}
-
-/// Application context (data associated with the app).
-#[derive(Clone)]
-pub enum AppContext {
-    /// Context of unregistered app.
-    Unregistered(Rc<Unregistered>),
-    /// Context of registered app.
-    Registered(Rc<Registered>),
-}
-
-#[allow(missing_docs)]
-pub struct Unregistered {
-    object_cache: ObjectCache,
-}
-
-#[allow(missing_docs)]
-pub struct Registered {
-    object_cache: ObjectCache,
-    app_id: String,
-    sym_enc_key: secretbox::Key,
-    access_container_info: AccessContInfo,
-    access_info: RefCell<HashMap<String, (MDataInfo, BTreeSet<Permission>)>>,
-}
-
-impl AppContext {
-    fn unregistered() -> Self {
-        AppContext::Unregistered(Rc::new(Unregistered { object_cache: ObjectCache::new() }))
-    }
-
-    fn registered(app_id: String,
-                  sym_enc_key: secretbox::Key,
-                  access_container_info: AccessContInfo)
-                  -> Self {
-        AppContext::Registered(Rc::new(Registered {
-            object_cache: ObjectCache::new(),
-            app_id: app_id,
-            sym_enc_key: sym_enc_key,
-            access_container_info: access_container_info,
-            access_info: RefCell::new(HashMap::new()),
-        }))
-    }
-
-    /// Object cache
-    pub fn object_cache(&self) -> &ObjectCache {
-        match *self {
-            AppContext::Unregistered(ref context) => &context.object_cache,
-            AppContext::Registered(ref context) => &context.object_cache,
-        }
-    }
-
-    /// Symmetric encryption/decryption key.
-    pub fn sym_enc_key(&self) -> Result<&secretbox::Key, AppError> {
-        Ok(&self.as_registered()?.sym_enc_key)
-    }
-
-    /// Refresh access info by fetching it from the network.
-    pub fn refresh_access_info(&self, client: &Client) -> Box<AppFuture<()>> {
-        let reg = fry!(self.as_registered()).clone();
-        refresh_access_info(reg, client)
-    }
-
-    /// Fetch mdata_info for the given container name.
-    pub fn get_container_mdata_info<T: Into<String>>(&self,
-                                                     client: &Client,
-                                                     name: T)
-                                                     -> Box<AppFuture<MDataInfo>> {
-        let reg = fry!(self.as_registered()).clone();
-        let name = name.into();
-
-        fetch_access_info(reg.clone(), client)
-            .and_then(move |_| {
-                let access_info = reg.access_info.borrow();
-                access_info.get(&name)
-                    .map(|&(ref mdata_info, _)| mdata_info.clone())
-                    .ok_or(AppError::NoSuchContainer)
-            })
-            .into_box()
-    }
-
-    /// Check the given permission for the given directory.
-    pub fn is_permitted<T: Into<String>>(&self,
-                                         client: &Client,
-                                         name: T,
-                                         permission: Permission)
-                                         -> Box<AppFuture<bool>> {
-        let reg = fry!(self.as_registered()).clone();
-        let name = name.into();
-
-        fetch_access_info(reg.clone(), client)
-            .and_then(move |_| {
-                let access_info = reg.access_info.borrow();
-                access_info.get(&name)
-                    .map(|&(_, ref permissions)| permissions.contains(&permission))
-                    .ok_or(AppError::NoSuchContainer)
-            })
-            .into_box()
-    }
-
-    fn as_registered(&self) -> Result<&Rc<Registered>, AppError> {
-        match *self {
-            AppContext::Registered(ref a) => Ok(a),
-            AppContext::Unregistered(_) => Err(AppError::OperationForbidden),
-        }
-    }
-}
-
-fn refresh_access_info(context: Rc<Registered>, client: &Client) -> Box<AppFuture<()>> {
-    let entry_key = {
-        let app_id_hash = sha256::hash(context.app_id.as_bytes()).0;
-        let nonce = context.access_container_info.nonce;
-
-        fry!(utils::symmetric_encrypt(&app_id_hash, &context.sym_enc_key, Some(&nonce)))
-    };
-
-    client.get_mdata_value(context.access_container_info.id,
-                         context.access_container_info.tag,
-                         entry_key)
-        .map_err(AppError::from)
-        .and_then(move |value| {
-            let encoded = utils::symmetric_decrypt(&value.content, &context.sym_enc_key)?;
-            let decoded = deserialise(&encoded)?;
-
-            *context.access_info.borrow_mut() = decoded;
-
-            Ok(())
-        })
-        .into_box()
-}
-
-fn fetch_access_info(context: Rc<Registered>, client: &Client) -> Box<AppFuture<()>> {
-    if context.access_info.borrow().is_empty() {
-        refresh_access_info(context, client)
-    } else {
-        future::ok(()).into_box()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use futures::Future;
-    use safe_core::{DIR_TAG, MDataInfo};
-    use safe_core::ipc::Permission;
-    use std::collections::HashMap;
-    use test_utils::{create_app_with_access, run};
-
-    #[test]
-    fn refresh_access_info() {
-        // Shared container
-        let container_info = unwrap!(MDataInfo::random_private(DIR_TAG));
-
-        let mut container_permissions = HashMap::new();
-        let _ = container_permissions.insert("_test".to_string(),
-                                             (container_info.clone(),
-                                              btree_set![Permission::Read, Permission::Insert]));
-
-        let app = create_app_with_access(container_permissions.clone(), false);
-
-        run(&app, move |client, context| {
-            let reg = unwrap!(context.as_registered()).clone();
-            assert!(reg.access_info.borrow().is_empty());
-
-            context.refresh_access_info(client)
-                .then(move |result| {
-                    unwrap!(result);
-                    let access_info = reg.access_info.borrow();
-                    assert_eq!(access_info.len(), 1);
-                    assert_eq!(unwrap!(access_info.get("_test")).0, container_info);
-                    assert_eq!(unwrap!(access_info.get("_test")).1,
-                               unwrap!(container_permissions.get("_test")).1);
-
-                    Ok(())
-                })
-        });
-    }
-
-    #[test]
-    fn get_container_mdata_info() {
-        // Shared container
-        let container_info = unwrap!(MDataInfo::random_private(DIR_TAG));
-        let cont_name = "_test".to_string();
-
-        let mut container_permissions = HashMap::new();
-        let _ = container_permissions.insert(cont_name.clone(),
-                                             (container_info.clone(),
-                                              btree_set![Permission::Read]));
-
-        let app = create_app_with_access(container_permissions, false);
-
-        run(&app, move |client, context| {
-            context.get_container_mdata_info(client, cont_name)
-                .then(move |res| {
-                    let info = unwrap!(res);
-                    assert_eq!(info, container_info);
-
-                    Ok(())
-                })
-        });
-    }
-
-    #[test]
-    fn is_permitted() {
-        // Shared container
-        let container_info = unwrap!(MDataInfo::random_private(DIR_TAG));
-        let cont_name = "_test".to_string();
-
-        let mut container_permissions = HashMap::new();
-        let _ = container_permissions.insert(cont_name.clone(),
-                                             (container_info.clone(),
-                                              btree_set![Permission::Read]));
-
-        let app = create_app_with_access(container_permissions, false);
-
-        run(&app, move |client, context| {
-            let f1 = context.is_permitted(client, cont_name.clone(), Permission::Read)
-                .then(move |res| {
-                    assert!(unwrap!(res));
-                    Ok(())
-                });
-
-            let f2 = context.is_permitted(client, cont_name.clone(), Permission::Insert)
-                .then(move |res| {
-                    assert!(!unwrap!(res));
-                    Ok(())
-                });
-
-            f1.join(f2).map(|_| ())
-        });
-
-    }
-}
+pub use app::{App, AppContext};
